Embold: vunerability

+++ src/include/rdb/faccbindev.h
@@ -14,7 +14,7 @@ class binaryDeviceAccessor : public FileAccessorInterface<T> {
   /**
    * @brief Posix File Descriptor
    */
-  ssize_t fd;
+  int fd;
 
  public:
   ~binaryDeviceAccessor();
